Add test enemy to levels
level.py

added create_enemies function to create a test Enemy object in a fixed
position

fixed a bug in the conditionals in floor_tiles and non_floor tiles in YSortCameraGroup
to use isinstance to check if a sprite is an Entity instead of a just a
Player. type() was causing the sprite.tiled_layer check to access an
attribute of Enemy which did not exist.

enemy.py

added an dummy input method which should be overwritten in child classes
if needed to prevent pygame.sprite.Group.update() throwing an error