Fix bug in Enemy class causing rapid movement speed
Entity Class
Add move_x and move_y methods.
move method now only updates positions and DOESN'T check for collisions

Enemy class
Fix bug in moveing upwards conditional statement where direction had to
be checked against +1 and not -1

Player class
move method in Player class fully overrides the parent method.
move method updates player position and checks for collisions

note: I think the rapid movement issue was caused by the collsion method
in entity.py. It appered that the enemy object was dectecting a
collision with itself, since it is a memeber of the obstacle sprites
group. I think the solution is to override the parent collision method
of Enemy and implement logic to check to prevent it checking for
collisions with itself.