Fix animation bug for enemy entity
bug was caused by the status instance variable of the enemy not being
update when its direction vector changed.

Fix: Added an inner function called update_status to the move method in
the Enemy class. The function is called whenever the direction is
changed by the move method. Also set initial status to "up" in the Enemy
class init method.