Fix bug in Enemy class
Enemy will now display the correct direction image when seeking
(pathing towards) the player.

Fixed by normalising the direction vector between the enemies current
position and the next tile in it's path. Because the direction wasn't
beging normalised the x and y components were not equal to 0 or 1
exclusivly when moveing in cartesian directions causing all conditionals
in the update status method to evaluate to False.