Fixed bug causing incorrect Enemy image displayed
fix bug causing the incorrect image for Enemy sprites to be displayed
during diagonal movement. Solution was to prevent A* considering tiles
diagonal to the current positon. This was achinved by changing
get_neighbours method in Level class to only return the cartesian
neighbours of the current tile.

get_neighbours method in Level class re-named to get_cartesian_neighbours