Fix bug in SpawnPoint class
fixed a bug in the SpawnPoint class causing entities to spawn in
incorrect positions. The instance variable pos in the Hitbox class
correctly scales the world coordinates by the SCALE constant in
settings.py. However this value was being overridden in the child
SpawnPoint class (just had the value of the argument pos with NO
scaling). This was causing SpawnPoint objects to be squished on the
screen and not scaled correctly to match the tiles (in this it was
squeezed by a factor of 1/5 i.e tiled coords (10,10) should be scaled to
(50,50) but they wern't and insted had the original tiled coords (10,10).

This bug fix has removed the need to use the ORIGINAL_TILE_SIZE constant
in the get_tile method in the Level class. This has also fixed the bug
concerning the enemy not moving since they get_tile method has been
reverted back to it's original logic.