Modify transition function in LevelHandler class
The prior commit fixed the issue with spawnpoints not being mapped
corretly from the Tiled level editor to the game. Therefore the ad-hoc
solution of drawing the initial frame of the new level without an offset
(regular_draw method in YSortCameraGroup class) is no longer required.
The call to this function (regular_draw) has been removed.

The method regular draw has not been removed from the YSortCameraGroup
class incase it is required in the future.