Fix bug when Entity collides with player
LevelHandler class

in the Player object constructor the Level objects obstacle_sprites
group is passed in as an argument to the groups parameter. This allows
the collision method for all entities to check for and handle collsions
with the player (when the entity is the one who collidied with the
static player).

Enitity class
both inner functions of the collision method (horizontal_collision and
vertical_collision will check if the sprite being collided with is of
type Observable. If true the sprites notify function will be called, to
notify all observers of the collsion.
This fixes a bug which prevented the Enemies movement behavior being
switched to attack mode when a moving enemy collided with a static
player.