Fix bug in Enemy class
fixed a bug in the Enemy class where the enemy entity would not
regognise the player if it was stationary and within the enemies circle
of aggression or attack when the enemy was in patrol mode.

To fix this bug the patrol method now checks to see if the player is
either in the circle of attack or aggression and changes the movement
behaviour appropriatly, every time the enemy moves whilst in patrol
mode.