Modify method in Enemy class
removed check to see if player is in the enemies circle of attack in the
patrol_mode method of the Enemy class. This is not required since upon
collision the enemies behaviour will be changed to attack_mode.
The check was also causing a collison bug.

This check may be required in the future if ranged attacks are added.