Fix bug in Level class
get_alive_enemies was returning enemies who's status was set to "dead"
the method now checks that the status of the enemies is not equal to
dead before returning it.

release method in enemyObjectPool now calls the reset method on the
enemy object before adding it to the free list.