Fix bug which prevented graphics rendering
custom_draw method in YSortCameraGroup class where not checking if
spirte objects in the list comprehensions for the floor_tiles and
non_floor_tiles variables had an attribute called "tiled_layer" before
trying to access it. Therefore since the new HealthObject objects not an
instance of the LivingEntity class the non_floor_tiles list
comprehention was trying to access a "tiled_layer" attribute which
doesn't exist for hitbox objects.